+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/works/ApiErrorsTest.scala
@@ -1,166 +1,184 @@
 package uk.ac.wellcome.platform.api.works
 
 import com.twitter.finagle.http.Status
+import com.twitter.finatra.http.EmbeddedHttpServer
 
 class ApiErrorsTest extends ApiWorksTestBase {
 
   it(
     "returns a BadRequest error when malformed query parameters are presented") {
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=penguin",
-      andExpect = Status.BadRequest,
-      withJsonBody = badRequest("pageSize: 'penguin' is not a valid Integer")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=penguin",
+        andExpect = Status.BadRequest,
+        withJsonBody = badRequest(apiPrefix, "pageSize: 'penguin' is not a valid Integer")
+      )
+    }
   }
 
   it("returns a NotFound error when requesting a work with a non-existent id") {
-    val badId = "non-existing-id"
-    server.httpGet(
-      path = s"/$apiPrefix/works/$badId",
-      andExpect = Status.NotFound,
-      withJsonBody = notFound(s"Work not found for identifier $badId")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      val badId = "non-existing-id"
+      server.httpGet(
+        path = s"/$apiPrefix/works/$badId",
+        andExpect = Status.NotFound,
+        withJsonBody = notFound(apiPrefix, s"Work not found for identifier $badId")
+      )
+    }
   }
 
   it(
     "returns a BadRequest error if the user asks for a page size just over the maximum") {
-    val pageSize = 101
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=$pageSize",
-      andExpect = Status.BadRequest,
-      withJsonBody =
-        badRequest(s"pageSize: [$pageSize] is not less than or equal to 100")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      val pageSize = 101
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=$pageSize",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, s"pageSize: [$pageSize] is not less than or equal to 100")
+      )
+    }
   }
 
   it(
     "returns a BadRequest error if the user asks for an overly large page size") {
-    val pageSize = 100000
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=$pageSize",
-      andExpect = Status.BadRequest,
-      withJsonBody =
-        badRequest(s"pageSize: [$pageSize] is not less than or equal to 100")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      val pageSize = 100000
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=$pageSize",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, s"pageSize: [$pageSize] is not less than or equal to 100")
+      )
+    }
   }
 
   it("returns a BadRequest error if the user asks for zero-length pages") {
-    val pageSize = 0
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=$pageSize",
-      andExpect = Status.BadRequest,
-      withJsonBody =
-        badRequest(s"pageSize: [$pageSize] is not greater than or equal to 1")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      val pageSize = 0
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=$pageSize",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, s"pageSize: [$pageSize] is not greater than or equal to 1")
+      )
+    }
   }
 
   it("returns a BadRequest error if the user asks for a negative page size") {
-    val pageSize = -50
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=$pageSize",
-      andExpect = Status.BadRequest,
-      withJsonBody =
-        badRequest(s"pageSize: [$pageSize] is not greater than or equal to 1")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      val pageSize = -50
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=$pageSize",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, s"pageSize: [$pageSize] is not greater than or equal to 1")
+      )
+    }
   }
 
   it("returns a BadRequest error if the user asks for page 0") {
-    server.httpGet(
-      path = s"/$apiPrefix/works?page=0",
-      andExpect = Status.BadRequest,
-      withJsonBody = badRequest("page: [0] is not greater than or equal to 1")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?page=0",
+        andExpect = Status.BadRequest,
+        withJsonBody = badRequest(apiPrefix, "page: [0] is not greater than or equal to 1")
+      )
+    }
   }
 
   it("returns a BadRequest error if the user asks for a page before 0") {
-    server.httpGet(
-      path = s"/$apiPrefix/works?page=-50",
-      andExpect = Status.BadRequest,
-      withJsonBody =
-        badRequest("page: [-50] is not greater than or equal to 1")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?page=-50",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, "page: [-50] is not greater than or equal to 1")
+      )
+    }
   }
 
   it("returns multiple errors if there's more than one invalid parameter") {
-    server.httpGet(
-      path = s"/$apiPrefix/works?pageSize=-60&page=-50",
-      andExpect = Status.BadRequest,
-      withJsonBody = badRequest(
-        "page: [-50] is not greater than or equal to 1, pageSize: [-60] is not greater than or equal to 1")
-    )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=-60&page=-50",
+        andExpect = Status.BadRequest,
+        withJsonBody = badRequest(apiPrefix,
+          "page: [-50] is not greater than or equal to 1, pageSize: [-60] is not greater than or equal to 1")
+      )
+    }
   }
 
   it("returns a Bad Request error if asked for an invalid include") {
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works?includes=foo",
         andExpect = Status.BadRequest,
-        withJsonBody = badRequest("includes: 'foo' is not a valid include")
+        withJsonBody = badRequest(apiPrefix, "includes: 'foo' is not a valid include")
       )
     }
   }
 
   it("returns a Bad Request error if asked for more than one invalid include") {
-    eventually {
-      server.httpGet(
-        path = s"/$apiPrefix/works?includes=foo,bar",
-        andExpect = Status.BadRequest,
-        withJsonBody =
-          badRequest("includes: 'foo', 'bar' are not valid includes")
-      )
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+        server.httpGet(
+          path = s"/$apiPrefix/works?includes=foo,bar",
+          andExpect = Status.BadRequest,
+          withJsonBody =
+            badRequest(apiPrefix, "includes: 'foo', 'bar' are not valid includes")
+        )
     }
   }
 
   it(
     "returns a Bad Request error if asked for a mixture of valid and invalid includes") {
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works?includes=foo,identifiers,bar",
         andExpect = Status.BadRequest,
         withJsonBody =
-          badRequest("includes: 'foo', 'bar' are not valid includes")
+          badRequest(apiPrefix, "includes: 'foo', 'bar' are not valid includes")
       )
     }
   }
 
   it(
     "returns a Bad Request error if asked for an invalid include on an individual work") {
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works/nfdn7wac?includes=foo",
         andExpect = Status.BadRequest,
-        withJsonBody = badRequest("includes: 'foo' is not a valid include")
+        withJsonBody = badRequest(apiPrefix, "includes: 'foo' is not a valid include")
       )
     }
   }
 
   it("returns Not Found if you look up a non-existent index") {
-    eventually {
-      server.httpGet(
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>      server.httpGet(
         path = s"/$apiPrefix/works?_index=foobarbaz",
         andExpect = Status.NotFound,
-        withJsonBody = notFound("There is no index foobarbaz")
+        withJsonBody = notFound(apiPrefix, "There is no index foobarbaz")
       )
     }
   }
 
   it("returns Not Found if you ask for a non-existent work") {
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works/xhu96f9j",
         andExpect = Status.NotFound,
-        withJsonBody = notFound("Work not found for identifier xhu96f9j")
+        withJsonBody = notFound(apiPrefix, "Work not found for identifier xhu96f9j")
       )
     }
   }
 
   it("returns Bad Request if you ask for a malformed identifier") {
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works/zd224ncv]",
         andExpect = Status.BadRequest,
         withJsonBody =
-          badRequest("Unrecognised character in identifier zd224ncv]")
+          badRequest(apiPrefix, "Unrecognised character in identifier zd224ncv]")
       )
     }
   }
@@ -172,7 +190,7 @@ class ApiErrorsTest extends ApiWorksTestBase {
     // Elasticsearch has a number of "private" indexes, which don't have
     // a canonicalId field to sort on.  Trying to query one of these will
     // trigger one such exception!
-    eventually {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
       server.httpGet(
         path = s"/$apiPrefix/works?_index=.watches",
         andExpect = Status.InternalServerError,
@@ -188,23 +206,26 @@ class ApiErrorsTest extends ApiWorksTestBase {
   }
 
   it(
-    "returns a Bad Request error if you try to page beyond the first 10000 works") {
-    val queries = List(
-      "page=10000",
-      "pageSize=100&page=101",
-      "page=126&pageSize=80"
-    )
-    queries.foreach { query =>
-      println(s"Testing query=$query")
-      eventually {
-        server.httpGet(
-          path = s"/$apiPrefix/works?$query",
-          andExpect = Status.BadRequest,
-          withJsonBody =
-            badRequest("Only the first 10000 works are available in the API.")
-        )
-      }
+    "returns a Bad Request error if you try to access the 10000th page") {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?page=10000",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, "Only the first 10000 works are available in the API.")
+      )
     }
-  }
+    }
 
+  it(
+    "returns a Bad Request error if you try to get the 101th page with 100 results per page") {
+    withApiFixtures(apiVersion = "v1") { case (apiPrefix, _, _, server: EmbeddedHttpServer) =>
+      server.httpGet(
+        path = s"/$apiPrefix/works?pageSize=100&page=101",
+        andExpect = Status.BadRequest,
+        withJsonBody =
+          badRequest(apiPrefix, "Only the first 10000 works are available in the API.")
+      )
+    }
+    }
 }
